requirements.txt
diffusers[torch]
numpy
pandas
Pillow
scikit-learn
torch
torchvision
transformers
